Create unique setting values for the three initial challenges and pass the setting model to the digit grouping view.

+++ app/src/main/java/speednumbers/mastersofmemory/com/domain/repository/IRepository.java
@@ -25,12 +25,12 @@ public interface IRepository {
         void getGameList(GetGamesCallback callback);
 
 
-        void getChallengeList(int gameKey, GetChallengesCallback callback);
+        void getChallengeList(long gameKey, GetChallengesCallback callback);
         boolean deleteChallenge(Challenge challenge);
         boolean insertChallenge(Challenge challenge);
 
 
-        void getSettingsList(int challengeKey, GetSettingsCallback callback);
+        void getSettingsList(long challengeKey, GetSettingsCallback callback);
         boolean updateSetting(Setting setting);
 
 }
